FirstActivity working
main activity is being changed to SecondActivity
This transfer is currently a mess. Also FirstActivity still needs to be
edited to reflect full game setup capabilities so it can be removed from
the SecondActivity. The SecondActivity will then get a copy of the game
object and modify it and update UI as neccessary.